[Operator] Better builder pod cleanup logic
The logic was inversed when dealing with error messages where it would
mark the build as errored when the pod couldn't be found, which is to be
expected since the builder reconciler is the one deleting the pod.

Changed the event so it dispatch it's intent of deleting the pod only
when it happens. There's a possibility that instead of writing the event
*before* it actually happens, we decide to move it to *after* it
happened.

The reasoning of writing it before is that the error will surface if it
doesn't come through, so the user will know it attempted to delete the
pod, and the failure was caused by that event, which improves the
storytelling a little bit, imo.