Operator: Add BuildSpec values
Early work on getting most of the values needed for a build out in a
spec so it can be defined in WorkspaceSpec. Instead of starting at the
top level, I'm starting with what's needed now at the Build's level and
working my way up. This is likely to change in the future as I figure
what's missing or misaligned.

The Registry will have to have an extension to some kind of vault. The
idea is to start with a K8s Secret and add more feature from there. It's
reasonable to think we'll need to support a variety of Vault so this
struct will need to be extensible somehow so we can add more vaults as
needs grow.