Operator: Introduce new specs for Workspace
This starts the process of defining what kind of information the
Workspace needs in order to deploy a project properly. This adds an
ImageSpec and a RegistrySpec (a version of those existed prior but their
purpose has changed).

The Workspace now inludes a Componnents field which represent all the
components that are needed (server, db, frontend == 3 components).

A Component inludes an ImageSpec that will be passed down to the Build.
The idea with this spec is that multiple components can each include an
ImageSpec and the Workspace reconciler will de-dupe those ImageSpec and
only builds the image that are needed.

Currently, all ImageSpec are going out to a build, but there's going to
be an option in the future to tell the workspace that an ImageSpec is
not buildable (don't want to build a mysql image every time...)

This is a WIP and the structure and implementation is subject to change.