[Operator] Implement Admission and Specs for Workspace
This is a bigger commit than anticipated because the workspace would
crash with a nil pointer dereference because the Workspace's Spec has a
pointer to a Tag.

This is on purpose as I want it to be an optional field that gets
populated by the admission webhook eventually. Since the admission
webhook is not fully working yet, we're going to have to deal with this
commit and as the feature gets more polish, the code will stabilize.

This commit also includes some changes to the spec as I'm trying to
push the information from the Workspace down to the builder.