Operator: Switch Image/Registry Spec
I am not certain as to how the structure of those spec should be, I
think this is the last change until I hit a roadbloack with that
particular layout.

ImageSpec now includes a bunch of pointer references which means these
fields can all be nil. Hopefully, this will make it possible to specify
images that needs to be built and pushed to their registry and images
that comes from a 3rd party that only needs to be pulled.