Operator: Workspace can be created from YAML
Changes in this PR makes it now possible to use the YAML in a workspace
and spin up instances that will map to the content of the Workspace
YAML. This is basically removing most of the hardcoded values and moving
to use values from the CRD's definition.

The Ingress is still somewhat hardcoded and most things that takes an
array now expects the array to have only 1 item in it. Happy path works
but the failures aren't properly propagated now.